BC-9229 - move consent stuff to user module
