+++ apps/server/src/modules/provisioning/service/tsp-provisioning.service.ts
@@ -3,14 +3,11 @@ import { Account, AccountSave, AccountService } from '@modules/account';
 import { Class, ClassFactory, ClassService, ClassSourceOptions } from '@modules/class';
 import { RoleService } from '@modules/role';
 import { School, SchoolService } from '@modules/school';
-import { UserDo, UserService } from '@modules/user';
+import { Consent, ParentConsent, UserConsent, UserDo, UserService } from '@modules/user';
 import { Injectable } from '@nestjs/common';
 import { TypeGuard } from '@shared/common/guards';
 import { NotFoundLoggableException } from '@shared/common/loggable-exception';
 import { RoleReference } from '@shared/domain/domainobject';
-import { Consent } from '@shared/domain/domainobject/consent';
-import { ParentConsent } from '@shared/domain/domainobject/parent-consent';
-import { UserConsent } from '@shared/domain/domainobject/user-consent';
 import { RoleName } from '@shared/domain/interface';
 import { ObjectId } from 'bson';
 import { ExternalClassDto, ExternalSchoolDto, ExternalUserDto, OauthDataDto, ProvisioningSystemDto } from '../dto';
