+++ apps/server/src/modules/user/domain/do/index.ts
@@ -1,2 +1,5 @@
+export { Consent } from './consent';
+export { ParentConsent } from './parent-consent';
+export { UserConsent } from './user-consent';
 export { UserSourceOptions } from './user-source-options.do';
 export { SecondarySchoolReference, UserDo } from './user.do';
