+++ apps/server/src/modules/user/index.ts
@@ -1,4 +1,4 @@
-export { UserDo, UserSourceOptions } from './domain';
+export { Consent, ParentConsent, UserConsent, UserDo, UserSourceOptions } from './domain';
 export { UserService } from './service';
 export { UserConfig } from './user.config';
 export { UserModule } from './user.module';
