+++ apps/server/src/modules/user/repo/index.ts
@@ -1,3 +1,6 @@
+export { ConsentEntity } from './mikro-orm/consent.entity';
+export { ParentConsentEntity } from './mikro-orm/parent-consent.entity';
+export { UserConsentEntity } from './mikro-orm/user-consent.entity';
 export { UserDORepo } from './mikro-orm/user-do.repo';
 export { User, UserProperties, UserSchoolEmbeddable } from './mikro-orm/user.entity';
 export { UserRepo } from './mikro-orm/user.repo';
