--- apps/server/src/shared/domain/entity/consent/index.ts
@@ -1,3 +1,0 @@
-export { ConsentEntity } from './consent.entity';
-export { UserConsentEntity } from './user-consent.entity';
-export { ParentConsentEntity } from './parent-consent.entity';
